requirements.txt
opencv-python
Pillow
gensim
pandas
sklearn